Add build constraints to build on macOS x86
