+++ pkg/plugins/runtime/runtime.go
@@ -1,9 +1,6 @@
 package runtime
 
 import (
-	"fmt"
-
-	"github.com/genuinetools/bpfd/proc"
 	"github.com/quarkslab/kdigger/pkg/bucket"
 )
 
@@ -15,13 +12,6 @@ const (
 var bucketAliases = []string{"runtimes", "rt"}
 
 type RuntimeBucket struct{}
-
-func (n RuntimeBucket) Run() (bucket.Results, error) {
-	runtime := proc.GetContainerRuntime(0, 0)
-	res := bucket.NewResults(bucketName)
-	res.AddComment(fmt.Sprintf("The container runtime seems to be %s.", runtime))
-	return *res, nil
-}
 
 func Register(b *bucket.Buckets) {
 	b.Register(bucket.Bucket{
