+++ pkg/plugins/runtime/runtime_darwin.go
@@ -1,0 +1,11 @@
+package runtime
+
+import (
+	"errors"
+
+	"github.com/quarkslab/kdigger/pkg/bucket"
+)
+
+func (n RuntimeBucket) Run() (bucket.Results, error) {
+	return bucket.Results{}, errors.New("runtime detection is not supported on macOS x86")
+}
