+++ pkg/plugins/syscalls/syscalls_darwin.go
@@ -1,0 +1,11 @@
+package syscalls
+
+import (
+	"errors"
+
+	"github.com/quarkslab/kdigger/pkg/bucket"
+)
+
+func (n SyscallsBucket) Run() (bucket.Results, error) {
+	return bucket.Results{}, errors.New("syscall scan is not supported on macOS x86")
+}
